Rename Utils.cs to FilePathUtils.cs, make the class name match
